Add necessary files for reproducing the bug
This commit contains the necessary content to reproduce a bug with using
lerna on a project with local dependencies. The `react` dependency in
package a is merely there to make sure lerna spends some time running
`npm install` in that package.

To reproduce the issue:

Run `lerna clean` (just to make sure, can be skipped on the first try)

Run `lerna bootstrap`

I'm reproducing the bug with lerna 2.0.0-rc.4.